Rename files named `component.js` to `index.js`
