+++ src/Heading/index.js
