+++ src/HexInput/index.js
@@ -1,6 +1,6 @@
 import Color from "@snejugal/color";
-import Field from "../Field/component";
-import Fields from "../Fields/component";
+import Field from "../Field";
+import Fields from "../Fields";
 import PropTypes from "prop-types";
 import React from "react";
 import localization from "../localization";
