+++ src/Hint/index.js
