+++ src/VariableEditor/index.js
@@ -1,20 +1,20 @@
 import "./styles.scss";
 
-import Button from "../Button/component";
-import Buttons from "../Buttons/component";
+import Button from "../Button";
+import Buttons from "../Buttons";
 import { ReactComponent as Check } from "./check.svg";
 import Color from "@snejugal/color";
-import Dialog from "../Dialog/component";
-import Heading from "../Heading/component";
-import HexInput from "../HexInput/component";
-import Hint from "../Hint/component";
-import HslInput from "../HslInput/component";
-import Palettes from "../Palettes/component";
+import Dialog from "../Dialog";
+import Heading from "../Heading";
+import HexInput from "../HexInput";
+import Hint from "../Hint";
+import HslInput from "../HslInput";
+import Palettes from "../Palettes";
 import PropTypes from "prop-types";
 import React from "react";
-import RgbInput from "../RgbInput/component";
-import Tabs from "../Tabs/component";
-import VariablePreview from "../VariablePreview/component";
+import RgbInput from "../RgbInput";
+import Tabs from "../Tabs";
+import VariablePreview from "../VariablePreview";
 import { defaultValues } from "../atthemeVariables";
 import localization from "../localization";
 import readFile from "../readFile";
