+++ src/Workspace/index.js
@@ -3,19 +3,19 @@ import "./styles.scss";
 import * as database from "../database/api";
 import { allVariablesAmount, defaultValues } from "../atthemeVariables";
 import Attheme from "attheme-js";
-import Button from "../Button/component";
-import Buttons from "../Buttons/component";
+import Button from "../Button";
+import Buttons from "../Buttons";
 import { ReactComponent as DownloadIcon } from "./download-icon.svg";
-import Field from "../Field/component";
-import Hint from "../Hint/component";
-import PaletteEditor from "../PaletteEditor/component";
+import Field from "../Field";
+import Hint from "../Hint";
+import PaletteEditor from "../PaletteEditor";
 import PropTypes from "prop-types";
 import React from "react";
-import ScriptRunner from "../ScriptRunner/component";
-import Snackbar from "../Snackbar/component";
-import Spinner from "../Spinner/component";
-import VariableEditor from "../VariableEditor/component";
-import Variables from "../Variables/component";
+import ScriptRunner from "../ScriptRunner";
+import Snackbar from "../Snackbar";
+import Spinner from "../Spinner";
+import VariableEditor from "../VariableEditor";
+import Variables from "../Variables";
 import download from "../download";
 import localization from "../localization";
 import prepareTheme from "../prepareTheme";
